Fix bug reported by Vadim with constants in the source code that are large than 2^63.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@9104 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/invariants/formula/ToRationalFormulaVisitor.java
@@ -110,7 +110,7 @@ public class ToRationalFormulaVisitor implements ToFormulaVisitor<CompoundState,
   private @Nullable RationalFormula evaluate(InvariantsFormula<CompoundState> pFormula, Map<? extends String, ? extends InvariantsFormula<CompoundState>> pEnvironment) {
     CompoundState value = pFormula.accept(this.evaluationVisitor, pEnvironment);
     if (value.isSingleton()) {
-      return this.rfmgr.makeNumber(value.getLowerBound().longValue());
+      return this.rfmgr.makeNumber(value.getLowerBound());
     }
     return null;
   }
